fix: 구조 변경
