+++ src/app/basicauction/[id]/BasicAuctionInfoContent.tsx
@@ -1,0 +1,12 @@
+interface BasicAuctionInfoContentProps {
+  id: number;
+  content: string;
+}
+
+function BasicAuctionInfoContent({ id, content }: BasicAuctionInfoContentProps) {
+  console.log(id); // 임시 뺄수도
+
+  return <div>{content}</div>;
+}
+
+export default BasicAuctionInfoContent;
